* main cleanups * make applied stresses work again * make normal vectors for segments consistent * make intersection operation in vm_token not overflow
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@2718 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ examples/main_benchmark.cpp
@@ -186,11 +186,11 @@ int main(int argc, char *argv[])
 
 	BenchmarkMicrostructure micro = S1 ;//getMicrostructure(std::string(argv[1])) ;
 	BenchmarkPhenomenon pheno = ELASTICITY ;//getPhenomenon(std::string(argv[2])) ;
-	int order = 1; //atoi(argv[3]) ;
+	int order = 1; //atof(argv[3]) ;
 	if(order < 1 || order > 2) { order = 1 ; }
 	double scale = 100.;//atof(argv[4]) ;
 	prop = 3;//atof(argv[5]) ;
-	int sampling = 1000;//atoi(argv[6]) ;
+	int sampling = 1000;//atof(argv[6]) ;
 	double length = getLength(micro) ;
 
 	double size = scale*length ;
